Update Problem_206.cs
1_2_3_4_5_6_7_8_9_0

number the digits from abcdefghi

-we know that i is 0, since to be square it must end in either 1,4,6,9,25, or 00.
-digital root must be 1,4,7,or 9

digital root = (1+2+3+4+5+6+7+8+9+a+b+c+d+e+f+g+h) = sum(unknowns) % 9 == 1,4,7,9


n = 100*k
sqrt(n) = sqrt(100*k) = 10*sqrt(k)


so k is also a square.
Thus, we are looking for a square of the form:

k = 1_2_3_4_5_6_7_8_9

it must be that the square root of this number ends in 3 or 7.

thus, want x such that x^2 = k

if k is 17 digits, we know that x must be 9 digits. So x can be found by considering all nine 
digit numbers such that the last digit is a 3 or a 7 and the preceding digits form a number
divisible by 4.


thus, consider all 8 digit multiples of 4 (call this y). Append 3 and 7 and see if we square that and
get the result we are looking for.

True
The y that we found is: 55875204
The x that we found is: 312203845556673849
The value of x squared is: 97471241180375457447384584160474801